requirements.txt
IProgress
ipywidgets
python-dotenv
# Run "conda install -c conda-forge libmagic" first on Macs using Apple silicon
python-magic
tqdm

google-cloud-aiplatform
google-cloud-discoveryengine

langchain-core
langchain-community
langchain-openai 
langchain-qdrant 
langchain-google-community
langchain-google-genai
langchain-google-vertexai

ragas
unstructured
unstructured[md]
